requirement.txt
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.0.74
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
Pillow==9.5.0
streamlit==1.25.0